# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.DomainMonitor:
    properties:
      alertConfigId:
        description: 告警配置
        type: integer
      checkInterval:
        description: 检查间隔(秒)
        type: integer
      createdAt:
        type: string
      domain:
        description: 域名
        type: string
      enableAlert:
        description: 是否启用告警
        type: boolean
      enableSSL:
        description: 是否启用SSL检查
        type: boolean
      id:
        type: integer
      lastCheck:
        description: 最后检查时间
        type: string
      nextCheck:
        description: 下次检查时间
        type: string
      responseThreshold:
        description: 响应时间阈值(ms)
        type: integer
      responseTime:
        description: 响应时间(ms)
        type: integer
      sslExpiry:
        description: SSL过期时间
        type: string
      sslExpiryDays:
        description: SSL过期提前告警天数
        type: integer
      sslValid:
        description: SSL是否有效
        type: boolean
      status:
        description: '状态: normal, abnormal, paused, unknown'
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 运维管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: iom API
  version: "1.0"
paths:
...
  /monitor/domains:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DomainMonitor'
            type: array
      summary: 获取域名监控列表
      tags:
      - DomainMonitor
    post:
      consumes:
      - application/json
      parameters:
      - description: 域名监控信息
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/model.DomainMonitor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DomainMonitor'
      summary: 创建域名监控
      tags:
      - DomainMonitor
  /monitor/domains/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 删除域名监控
      tags:
      - DomainMonitor
    get:
      consumes:
      - application/json
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DomainMonitor'
      summary: 获取域名监控详情
      tags:
      - DomainMonitor
    put:
      consumes:
      - application/json
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: integer
      - description: 域名监控信息
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/model.DomainMonitor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DomainMonitor'
      summary: 更新域名监控
      tags:
      - DomainMonitor
  /monitor/domains/{id}/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: 域名ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DomainMonitor'
      summary: 立即检查域名
      tags:
      - DomainMonitor
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"